app/gateway/auth.py
import hashlib
from datetime import datetime, timedelta
from typing import Optional
from jose import JWTError, jwt
from passlib.context import CryptContext
import os

# Configurações JWT
SECRET_KEY = os.getenv("SECRET_KEY", "fiapx_super_secret_key_change_me")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60 * 24 # 1 dia

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

def verify_password(plain_password, hashed_password):
    # Pre-hash com SHA256 para evitar o limite de 72 bytes do bcrypt
    pwd_hash = hashlib.sha256(plain_password.encode()).hexdigest()
    return pwd_context.verify(pwd_hash, hashed_password)

def get_password_hash(password):
    # Pre-hash com SHA256 para evitar o limite de 72 bytes do bcrypt
    pwd_hash = hashlib.sha256(password.encode()).hexdigest()
    return pwd_context.hash(pwd_hash)

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt